requirements.txt
requests>=2.31.0
mcp>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
openai>=1.12.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.3
browser-use>=0.1.41
